# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/booking/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get all cars",
                "operationId": "f9280585c6c400fa699246fcd05a275a",
                "responses": {
                    "200": {
                        "description": "List of all cars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingCar"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Add a new car booking",
                "operationId": "e6ff6922c596d14ad7db64fd48770f10",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingCar"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Car booking added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingCar"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/booking/cars/{id}": {
            "put": {
                "tags": [
                    "Cars"
                ],
                "summary": "Update a car booking",
                "operationId": "e14280f30238625b22a43b5692e772d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingCar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car booking updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingCar"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car booking not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Cancel a car booking",
                "operationId": "6f60d96b5e1c73e1446a6bf7ac0506b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car booking canceled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingCar"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car booking not found"
                    }
                }
            }
        },
        "/api/booking/flights": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get list of all flights",
                "description": "Returns a list of all flights",
                "operationId": "getAllFlights",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingFlight"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Add a new flight booking",
                "description": "Creates a new flight booking",
                "operationId": "createFlightBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingFlight"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingFlight"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/booking/flights/{id}": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Get a flight by ID",
                "description": "Returns a specific flight by its ID",
                "operationId": "getFlightById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingFlight"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Flights"
                ],
                "summary": "Cancel a flight booking",
                "description": "Cancels a flight booking by ID",
                "operationId": "cancelFlightBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking canceled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingFlight"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Flight not found"
                    }
                }
            }
        },
        "/api/booking/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get all hotels",
                "operationId": "b190f12b8ba653c82276be8f6ed73962",
                "responses": {
                    "200": {
                        "description": "List of all hotels",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingHotel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Add a new hotel booking",
                "operationId": "e9abd4a3d968e6c41b594dd2b9022129",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingHotel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel booking added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingHotel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/booking/hotels/{id}": {
            "put": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Update a hotel booking",
                "operationId": "e6064d80e6fb6a16c67a3af481a9f0bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingHotel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel booking updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingHotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel booking not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Cancel a hotel booking",
                "operationId": "edd080b505aab34e47ca1ac008ec66dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel booking canceled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingHotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel booking not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookingFlight": {
                "title": "BookingFlight",
                "required": [
                    "UserID",
                    "AirlineName",
                    "DepartureAirport",
                    "ArrivalAirport",
                    "DepartureTime",
                    "ArrivalTime",
                    "Price",
                    "StartDate",
                    "EndDate"
                ],
                "properties": {
                    "UserID": {
                        "description": "The ID of the user",
                        "type": "integer"
                    },
                    "AirlineName": {
                        "description": "Name of the airline",
                        "type": "string"
                    },
                    "DepartureAirport": {
                        "description": "Departure airport",
                        "type": "string"
                    },
                    "ArrivalAirport": {
                        "description": "Arrival airport",
                        "type": "string"
                    },
                    "DepartureTime": {
                        "description": "Departure time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "ArrivalTime": {
                        "description": "Arrival time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "Price": {
                        "description": "Price of the flight",
                        "type": "number",
                        "format": "float"
                    },
                    "StartDate": {
                        "description": "Booking start date",
                        "type": "string",
                        "format": "date"
                    },
                    "EndDate": {
                        "description": "Booking end date",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "BookingCar": {
                "title": "BookingCar",
                "required": [
                    "UserID",
                    "CarModel",
                    "SeatNumber",
                    "Location",
                    "PricePerDay",
                    "Availability",
                    "StartDate",
                    "EndDate"
                ],
                "properties": {
                    "UserID": {
                        "type": "integer"
                    },
                    "CarModel": {
                        "type": "string"
                    },
                    "SeatNumber": {
                        "type": "integer"
                    },
                    "Location": {
                        "type": "string"
                    },
                    "PricePerDay": {
                        "type": "number",
                        "format": "float"
                    },
                    "Availability": {
                        "type": "boolean"
                    },
                    "StartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "EndDate": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "BookingHotel": {
                "title": "BookingHotel",
                "required": [
                    "UserID",
                    "HotelName",
                    "Location",
                    "RoomType",
                    "PricePerNight",
                    "Availability",
                    "StartDate",
                    "EndDate"
                ],
                "properties": {
                    "UserID": {
                        "type": "integer"
                    },
                    "HotelName": {
                        "type": "string"
                    },
                    "Location": {
                        "type": "string"
                    },
                    "RoomType": {
                        "type": "string"
                    },
                    "PricePerNight": {
                        "type": "number",
                        "format": "float"
                    },
                    "Availability": {
                        "type": "boolean"
                    },
                    "StartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "EndDate": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Flights",
            "description": "Flights"
        }
    ]
}